requirements.txt
opencv-python
numpy
pygetwindow
pyautogui
Pillow
openai
python-dotenv
